Correct definitions

+++ meshkernel/py_structures.py
@@ -756,7 +756,7 @@ class GriddedSamples:
             self.values: ndarray = np.asarray(values, dtype=np.float32)
         elif values.dtype == np.int64:
             self.value_type: int = InterpolationValues.INT
-            self.values: ndarray = np.asarray(values, dtype=np.int64)
+            self.values: ndarray = np.asarray(values, dtype=np.int32)
         elif values.dtype == np.float64:
             self.value_type: int = InterpolationValues.DOUBLE
             self.values: ndarray = np.asarray(values, dtype=np.float64)
